EC2 initial cfn stack template
This cloudformation stack defines an EC2 resource to be deployed through AWS CLI.

The template defines a security group to allow for tcp connection over port 80. Once the server is deployed, it updates the system software.

TODO: extend the cloudformation stack to allow for SSH connection and edit the on start commands on the server.